時間割表をモーダルからページ内へ移動
